feat: simplify port forwarding with automatic preview URL routing
- Add handleSandboxRequest() to automatically route preview URLs
- Support all localhost patterns (127.0.0.1, ::1, etc.)
- Reduce user code from 40+ lines to just 3-5 lines
- Move Sandbox class to separate file for better organization
- Export client response types for TypeScript support
- Add comprehensive port forwarding documentation to README

BREAKING CHANGE: Refactored exports - Sandbox class now imported from main package
